backend/requirements.txt
fastapi
uvicorn
openai
moviepy
python-dotenv
requests
boto3